requirements.txt
amqp==2.2.2
appnope==0.1.0
async-timeout==2.0.1
attrs==17.4.0
awscli==1.15.4
backcall==0.1.0
billiard==3.5.0.3
botocore==1.10.4
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.7
decorator==4.2.1
Django==2.0.4
django-cors-headers==2.2.0
django-environ==0.4.4
django-extensions==2.0.6
django-model-utils==3.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docutils==0.14
flake8==3.5.0
gevent==1.2.2
gunicorn==19.7.1
idna==2.6
idna-ssl==1.0.1
ipdb==0.11
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.11.1
jmespath==0.9.3
kombu==4.1.0
mccabe==0.6.1
more-itertools==4.1.0
multidict==4.2.0
omdb==0.9.1
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.4
ptyprocess==0.5.2
py==1.5.3
pyasn1==0.4.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.6.1
pytest==3.5.0
pytest-mock==1.8.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.11.0
tmdbsimple==2.1.0
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
yarl==1.1.1
